+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.4
 ---------------
 - Add XRedirectUrl JS global var for soft redirection when not an AJAX request in PreparePHP_SELF.fnc.php & warehouse.js
+- JS Fix double HTML character encoding, use jQuery text instead of jQuery html in warehouse.js
 
 Changes in 11.3.2
 -----------------